refactor: update imports after moving time-related functions
